Update guide_mailman-3.rst
installing urllib3==1.26.16 allows for it to work with CentOS7 (otherwise there is a problem with OpenSSL<1.1.1+
despite flufl.i18n>3.2 being given as a dependency, 5.0 doesn't seem to be recognised. Reverting to 3.2 fixes this, I have not tested higher versions. 
I also upgraded setuptools which may have contributed to everything working.